fix: Исправлен экран выбора валюты, переписан с overlay на popover
