+++ Expense Tracker/Services/BankAccountManager.swift
@@ -1,10 +1,5 @@
 //
 //  BankAccountManager.swift
-//  Expense Tracker
-//
-//  Created by user on 26.06.2025.
-//
-
 import Foundation
 
 protocol BankAccountStorage {
