+++ Expense Tracker/Services/BankAccountsService.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-final class BankAccountsService {
+actor BankAccountsService {
     private var account = BankAccount(
         id: 1,
         userId: 1,
